oci: more grep'able interface name
`git grep -wi store` is not nearly useful enough. Taking steps for
readability.
